+++ src/app/(authenticated)/presensi-siswa/rekap/page.tsx
@@ -1,19 +1,16 @@
 "use client";
 
 import BaseLoading from "@/components/base-loading";
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
 import { Separator } from "@/components/ui/separator";
-import { Skeleton } from "@/components/ui/skeleton";
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
-import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip";
-import { useMonthlyApelAttendanceQuery, useStudentApelHistoryQuery } from "@/queries/useApelAttendanceQuery";
+import { useMonthlyApelAttendanceQuery } from "@/queries/useApelAttendanceQuery";
 import { useRombelsQuery } from "@/queries/useRombelQuery";
-import { IconArrowLeft, IconChevronRight, IconUsers } from "@tabler/icons-react";
-import { Clock, Layers, User } from "lucide-react";
+import { IconCalendarMonth, IconUsers } from "@tabler/icons-react";
+import { User } from "lucide-react";
+import Link from "next/link";
+import { useSearchParams } from "next/navigation";
 import { useMemo, useState } from "react";
 import MonthYearSelector from "../../rekap/bulanan/components/month-year-selector";
 
@@ -32,27 +29,16 @@ const monthsList = [
 	{ value: 12, label: "Desember" },
 ];
 
-export default function PresensiSiswaRekapPage() {
+export default function PresensiSiswaRekapMainPage() {
+	const searchParams = useSearchParams();
 	const currentDate = new Date();
-	const [month, setMonth] = useState(currentDate.getMonth() + 1);
-	const [year, setYear] = useState(currentDate.getFullYear());
-	const [selectedRombel, setSelectedRombel] = useState<string>("");
-	const [selectedStudentId, setSelectedStudentId] = useState<string | null>(null);
-	const [viewType, setViewType] = useState<"ringkasan" | "matriks">("ringkasan");
+	const initialMonth = Number(searchParams.get("month")) || currentDate.getMonth() + 1;
+	const initialYear = Number(searchParams.get("year")) || currentDate.getFullYear();
+
+	const [month, setMonth] = useState(initialMonth);
+	const [year, setYear] = useState(initialYear);
 
 	const { data: rombels, isLoading: isRombelsLoading } = useRombelsQuery();
-
-	const { data: monthlyData, isLoading: isMonthlyLoading } = useMonthlyApelAttendanceQuery({
-		rombel_id: selectedRombel,
-		month,
-		year,
-	});
-
-	const { data: studentHistoryData, isLoading: isStudentHistoryLoading } = useStudentApelHistoryQuery({
-		student_id: selectedStudentId || "",
-		month,
-		year,
-	});
 
 	// Group rombels by grade level (tingkat_kelas)
 	const groupedRombels = useMemo(() => {
@@ -74,74 +60,6 @@ export default function PresensiSiswaRekapPage() {
 			.sort((a, b) => a - b);
 	}, [groupedRombels]);
 
-	// Days calculation for matrix
-	const daysInMonth = new Date(year, month, 0).getDate();
-	const daysArray = Array.from({ length: daysInMonth }, (_, i) => i + 1);
-
-	const studentAttendanceMap = monthlyData?.attendances ?? {};
-	const studentKeys = Object.keys(studentAttendanceMap);
-
-	const selectedRombelObj = rombels?.find((r) => r.id === selectedRombel);
-
-	// Calculate student summary stats
-	const studentStatsList = useMemo(() => {
-		return studentKeys.map((stId) => {
-			const records = studentAttendanceMap[stId] || [];
-			const firstStudent = records[0]?.student;
-			const studentName = firstStudent?.fullname || stId;
-			const nipd = firstStudent?.nipd || "-";
-			const photo = firstStudent?.photo || "";
-
-			let countH = 0;
-			let countS = 0;
-			let countI = 0;
-			let countA = 0;
-
-			const dayStatusMap: Record<number, string> = {};
-
-			records.forEach((rec) => {
-				const dateObj = new Date(rec.attendance_date);
-				const dayNum = dateObj.getDate();
-				const status = (rec.attendance_status || "h").toLowerCase();
-				dayStatusMap[dayNum] = status;
-
-				if (status === "h") countH++;
-				else if (status === "s") countS++;
-				else if (status === "i") countI++;
-				else if (status === "a") countA++;
-			});
-
-			const totalDays = countH + countS + countI + countA;
-			const rate = totalDays > 0 ? Math.round((countH / totalDays) * 100) : 0;
-
-			return {
-				stId,
-				studentName,
-				nipd,
-				photo,
-				countH,
-				countS,
-				countI,
-				countA,
-				totalDays,
-				rate,
-				dayStatusMap,
-				records,
-			};
-		});
-	}, [studentKeys, studentAttendanceMap]);
-
-	const selectedStudentDetail = useMemo(() => {
-		if (!selectedStudentId) return null;
-		return studentStatsList.find((s) => s.stId === selectedStudentId) || null;
-	}, [selectedStudentId, studentStatsList]);
-
-	const getAttendanceRateBadge = (rate: number) => {
-		if (rate >= 90) return <Badge className="bg-emerald-500 hover:bg-emerald-600 font-bold">{rate}%</Badge>;
-		if (rate >= 75) return <Badge className="bg-amber-500 hover:bg-amber-600 font-bold">{rate}%</Badge>;
-		return <Badge variant="destructive" className="font-bold">{rate}%</Badge>;
-	};
-
 	const monthName = monthsList.find((m) => m.value === month)?.label;
 
 	return (
@@ -157,239 +75,28 @@ export default function PresensiSiswaRekapPage() {
 							Rekap kehadiran siswa per bulan berdasarkan kelas dan presensi apel pagi
 						</p>
 					</div>
-
-					{selectedStudentId ? (
-						<Button
-							variant="outline"
-							onClick={() => setSelectedStudentId(null)}
-							className="mt-4 md:mt-0 font-medium"
-						>
-							<IconArrowLeft className="mr-2 w-4 h-4" /> Kembali ke {monthlyData?.rombel || selectedRombelObj?.nama || "Daftar Siswa"}
-						</Button>
-					) : selectedRombel ? (
-						<Button
-							variant="outline"
-							onClick={() => setSelectedRombel("")}
-							className="mt-4 md:mt-0 font-medium"
-						>
-							<IconArrowLeft className="mr-2 w-4 h-4" /> Kembali ke Rekap Bulanan
-						</Button>
-					) : null}
 				</div>
 
 				<Separator className="my-4" />
 
 				{/* Filter Row matching /rekap/bulanan */}
-				{!selectedStudentId && (
-					<>
-						<div className="grid grid-cols-12 gap-4 items-center">
-							<div className="col-span-12 md:col-span-6">
-								<MonthYearSelector
-									month={month}
-									year={year}
-									onMonthChange={setMonth}
-									onYearChange={setYear}
-								/>
-							</div>
+				<div className="grid grid-cols-12 gap-4 items-center">
+					<div className="col-span-12 md:col-span-6">
+						<MonthYearSelector
+							month={month}
+							year={year}
+							onMonthChange={setMonth}
+							onYearChange={setYear}
+						/>
+					</div>
+				</div>
 
-							{selectedRombel && (
-								<div className="col-span-12 md:col-span-6 flex justify-end items-end gap-3">
-									<div className="w-56 space-y-1">
-										<label htmlFor="view-type-select" className="text-xs font-medium text-muted-foreground flex items-center gap-1">
-											<Layers className="w-3.5 h-3.5" /> Tampilan Detail
-										</label>
-										<Select value={viewType} onValueChange={(val: "ringkasan" | "matriks") => setViewType(val)}>
-											<SelectTrigger id="view-type-select" className="h-9 text-xs">
-												<SelectValue placeholder="Pilih Tampilan" />
-											</SelectTrigger>
-											<SelectContent>
-												<SelectItem value="ringkasan">Daftar Ringkasan Siswa</SelectItem>
-												<SelectItem value="matriks">Matriks Presensi Harian (1-{daysInMonth})</SelectItem>
-											</SelectContent>
-										</Select>
-									</div>
-								</div>
-							)}
-						</div>
-
-						<Separator className="my-4" />
-					</>
-				)}
+				<Separator className="my-4" />
 
 				{/* CONTENT AREA */}
 				{isRombelsLoading ? (
 					<BaseLoading />
-				) : selectedStudentId ? (
-					/* INDIVIDUAL STUDENT DETAIL VIEW (FULL MATCH TO /rekap/bulanan/[rombelId]/siswa/[studentId]) */
-					<div className="space-y-6">
-						{/* Header Banner */}
-						<div className="flex flex-col md:flex-row items-start md:items-center justify-between gap-4 bg-card p-6 rounded-xl border shadow-xs">
-							<div className="flex items-center gap-4">
-								<Avatar className="w-16 h-16 border-2 border-primary/20 shadow-xs">
-									<AvatarImage src={selectedStudentDetail?.photo} alt={selectedStudentDetail?.studentName || "Siswa"} />
-									<AvatarFallback className="text-2xl font-bold bg-primary/10 text-primary">
-										{(selectedStudentDetail?.studentName || "S").charAt(0).toUpperCase()}
-									</AvatarFallback>
-								</Avatar>
-								<div className="space-y-1">
-									<h2 className="font-bold text-2xl tracking-tight">
-										{selectedStudentDetail?.studentName || "Detail Siswa"}
-									</h2>
-									<p className="text-muted-foreground text-sm font-medium">
-										NIS: {selectedStudentDetail?.nipd || "-"} | {monthlyData?.rombel || selectedRombelObj?.nama}
-									</p>
-									<p className="text-muted-foreground text-xs">
-										Periode Presensi Apel: {monthName} {year}
-									</p>
-								</div>
-							</div>
-
-							<div className="flex items-center gap-2">
-								{getAttendanceRateBadge(selectedStudentDetail?.rate || 0)}
-							</div>
-						</div>
-
-						{/* 6 Metric Stat Cards */}
-						<div className="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-6 gap-3">
-							<Card className="shadow-2xs">
-								<CardContent className="p-4 text-center">
-									<div className="font-bold text-emerald-600 text-2xl">
-										{selectedStudentDetail?.countH || 0}
-									</div>
-									<div className="text-muted-foreground text-xs mt-0.5">Hadir</div>
-								</CardContent>
-							</Card>
-							<Card className="shadow-2xs">
-								<CardContent className="p-4 text-center">
-									<div className="font-bold text-sky-600 text-2xl">
-										{selectedStudentDetail?.countS || 0}
-									</div>
-									<div className="text-muted-foreground text-xs mt-0.5">Sakit</div>
-								</CardContent>
-							</Card>
-							<Card className="shadow-2xs">
-								<CardContent className="p-4 text-center">
-									<div className="font-bold text-indigo-600 text-2xl">
-										{selectedStudentDetail?.countI || 0}
-									</div>
-									<div className="text-muted-foreground text-xs mt-0.5">Izin</div>
-								</CardContent>
-							</Card>
-							<Card className="shadow-2xs">
-								<CardContent className="p-4 text-center">
-									<div className="font-bold text-amber-600 text-2xl">
-										{selectedStudentDetail?.countA || 0}
-									</div>
-									<div className="text-muted-foreground text-xs mt-0.5">Alpa</div>
-								</CardContent>
-							</Card>
-							<Card className="shadow-2xs">
-								<CardContent className="p-4 text-center">
-									<div className="font-bold text-foreground text-2xl">
-										{selectedStudentDetail?.totalDays || 0}
-									</div>
-									<div className="text-muted-foreground text-xs mt-0.5">Total Sesi</div>
-								</CardContent>
-							</Card>
-							<Card className="shadow-2xs">
-								<CardContent className="p-4 text-center">
-									<div className="font-bold text-emerald-600 text-2xl">
-										{selectedStudentDetail?.rate || 0}%
-									</div>
-									<div className="text-muted-foreground text-xs mt-0.5">Persentase</div>
-								</CardContent>
-							</Card>
-						</div>
-
-						{/* Kalender Kehadiran Card (EXACT MATCH TO /rekap/bulanan/[rombelId]/siswa/[studentId]) */}
-						<Card className="shadow-xs">
-							<CardHeader>
-								<CardTitle className="text-xl">Kalender Kehadiran Apel</CardTitle>
-								<CardDescription className="text-xs">
-									Riwayat visual kehadiran apel pagi siswa selama bulan {monthName} {year}
-								</CardDescription>
-							</CardHeader>
-							<CardContent>
-								<AttendanceCalendar
-									month={month}
-									year={year}
-									dayStatusMap={selectedStudentDetail?.dayStatusMap || {}}
-								/>
-							</CardContent>
-						</Card>
-
-						{/* History Log Table */}
-						<Card className="shadow-xs overflow-hidden">
-							<CardHeader className="bg-muted/30 pb-3">
-								<CardTitle className="text-lg">Rincian Log Presensi Apel</CardTitle>
-								<CardDescription className="text-xs">
-									Log pemindaian dan jam presensi apel per tanggal di bulan {monthName} {year}
-								</CardDescription>
-							</CardHeader>
-							<CardContent className="p-0">
-								{isStudentHistoryLoading ? (
-									<div className="p-6 space-y-3">
-										<Skeleton className="w-full h-10" />
-										<Skeleton className="w-full h-10" />
-									</div>
-								) : !selectedStudentDetail?.records || selectedStudentDetail.records.length === 0 ? (
-									<div className="py-12 text-center text-muted-foreground text-sm italic">
-										Belum ada rincian presensi apel untuk siswa ini di bulan {monthName} {year}.
-									</div>
-								) : (
-									<div className="overflow-x-auto">
-										<Table>
-											<TableHeader>
-												<TableRow className="bg-muted/50 border-b">
-													<TableHead className="w-12 text-center font-bold px-3 text-xs">No</TableHead>
-													<TableHead className="font-bold text-xs px-4">Tanggal Presensi</TableHead>
-													<TableHead className="text-center font-bold text-xs px-3">Jam Masuk / Scan</TableHead>
-													<TableHead className="text-center font-bold text-xs px-3">Status</TableHead>
-												</TableRow>
-											</TableHeader>
-											<TableBody>
-												{selectedStudentDetail.records.map((rec, idx) => {
-													const dateObj = new Date(rec.attendance_date);
-													const formattedDate = dateObj.toLocaleDateString("id-ID", {
-														weekday: "long",
-														day: "numeric",
-														month: "long",
-														year: "numeric",
-													});
-													const formattedTime = dateObj.toLocaleTimeString("id-ID", {
-														hour: "2-digit",
-														minute: "2-digit",
-													});
-
-													const status = (rec.attendance_status || "h").toLowerCase();
-													let statusBadge = <Badge className="bg-emerald-500 text-white font-bold">Hadir</Badge>;
-
-													if (status === "s") statusBadge = <Badge className="bg-sky-500 text-white font-bold">Sakit</Badge>;
-													else if (status === "i") statusBadge = <Badge className="bg-indigo-500 text-white font-bold">Izin</Badge>;
-													else if (status === "a") statusBadge = <Badge className="bg-amber-500 text-white font-bold">Alpa</Badge>;
-
-													return (
-														<TableRow key={rec.id || idx} className="hover:bg-muted/40 transition-colors border-b">
-															<TableCell className="text-center text-xs text-muted-foreground px-3">{idx + 1}</TableCell>
-															<TableCell className="font-medium text-xs px-4">{formattedDate}</TableCell>
-															<TableCell className="text-center text-xs font-mono px-3">
-																<span className="inline-flex items-center gap-1">
-																	<Clock className="w-3.5 h-3.5 text-muted-foreground" /> {formattedTime} WIB
-																</span>
-															</TableCell>
-															<TableCell className="text-center px-3">{statusBadge}</TableCell>
-														</TableRow>
-													);
-												})}
-											</TableBody>
-										</Table>
-									</div>
-								)}
-							</CardContent>
-						</Card>
-					</div>
-				) : !selectedRombel ? (
-					/* MAIN VIEW: CLASS CARDS GRID (FULL MATCH TO /rekap/bulanan) */
+				) : (
 					<div className="space-y-6">
 						{/* Summary Banner */}
 						<div className="flex flex-wrap justify-between items-center gap-4 bg-muted p-4 rounded-lg">
@@ -407,248 +114,17 @@ export default function PresensiSiswaRekapPage() {
 							<div key={grade} className="space-y-4">
 								<h2 className="font-medium text-xl">Kelas {grade}</h2>
 								<div className="gap-4 grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3">
-									{groupedRombels[grade]?.map((rombel) => {
-										const waliName =
-											typeof rombel.wali_kelas === "object"
-												? rombel.wali_kelas?.fullname
-												: typeof rombel.wali_kelas === "string"
-													? rombel.wali_kelas
-													: "Wali Kelas";
-
-										const jurusanNama =
-											typeof rombel.jurusan === "object"
-												? rombel.jurusan?.nama || rombel.jurusan?.kode || "Presensi Apel Pagi"
-												: rombel.jurusan || "Presensi Apel Pagi";
-
-										return (
-											<Card
-												key={rombel.id}
-												onClick={() => setSelectedRombel(rombel.id)}
-												className="hover:shadow-md h-full hover:scale-95 transition duration-300 cursor-pointer border"
-											>
-												<CardHeader className="pb-2">
-													<CardTitle className="flex justify-between items-center text-lg">
-														<span>{rombel.nama}</span>
-														<Badge variant="secondary" className="text-xs font-normal">
-															Tingkat {rombel.tingkat_kelas}
-														</Badge>
-													</CardTitle>
-													<CardDescription>{jurusanNama}</CardDescription>
-												</CardHeader>
-												<CardContent>
-													<div className="space-y-2">
-														<div className="flex items-center gap-2 text-muted-foreground text-xs">
-															<User className="w-4 h-4" />
-															<span>{waliName}</span>
-														</div>
-														<div className="flex items-center gap-2 text-muted-foreground text-xs">
-															<IconUsers className="w-4 h-4" />
-															<span>Klik untuk buka rekap detail siswa</span>
-														</div>
-													</div>
-												</CardContent>
-											</Card>
-										);
-									})}
+									{groupedRombels[grade]?.map((rombel) => (
+										<RombelApelCard
+											key={rombel.id}
+											rombel={rombel}
+											month={month}
+											year={year}
+										/>
+									))}
 								</div>
 							</div>
 						))}
-					</div>
-				) : (
-					/* CLASS STUDENT TABLE VIEW (FULL MATCH TO /rekap/bulanan/[rombelId]) */
-					<div className="space-y-5">
-						<div className="flex flex-row justify-between items-start gap-4">
-							<div className="space-y-1">
-								<h2 className="font-semibold text-2xl tracking-tight text-primary">
-									{monthlyData?.rombel || selectedRombelObj?.nama || "Rombel"}
-								</h2>
-								<p className="text-muted-foreground text-sm font-medium">
-									{(typeof selectedRombelObj?.jurusan === "object"
-										? selectedRombelObj?.jurusan?.nama || selectedRombelObj?.jurusan?.kode
-										: selectedRombelObj?.jurusan) || "Presensi Apel Pagi"}{" "}
-									| {monthName} {year}
-								</p>
-							</div>
-						</div>
-
-						{/* Summary Stats Badges */}
-						<div className="flex flex-wrap gap-4 mb-2">
-							<Badge variant="outline" className="px-3 py-1 text-sm">
-								Total Siswa: {studentKeys.length}
-							</Badge>
-							<Badge variant="outline" className="px-3 py-1 text-sm">
-								Total Hari Sesi: {daysInMonth} Hari
-							</Badge>
-						</div>
-
-						{/* Students Table matching /rekap/bulanan/[rombelId] */}
-						<Card className="border-none shadow-md overflow-hidden bg-card/50 backdrop-blur-xs">
-							<CardHeader className="bg-muted/30">
-								<CardTitle className="text-xl">Daftar Siswa</CardTitle>
-								<CardDescription>
-									Klik pada nama siswa untuk melihat detail kehadiran individual
-								</CardDescription>
-							</CardHeader>
-							<CardContent className="p-0">
-								{isMonthlyLoading ? (
-									<div className="p-6 space-y-3">
-										<Skeleton className="w-full h-12" />
-										<Skeleton className="w-full h-12" />
-										<Skeleton className="w-full h-12" />
-									</div>
-								) : studentStatsList.length === 0 ? (
-									<div className="py-12 text-center text-muted-foreground text-sm italic">
-										Tidak ada data presensi siswa untuk kelas ini pada bulan {monthName} {year}.
-									</div>
-								) : viewType === "ringkasan" ? (
-									/* SUMMARY STUDENT TABLE (CLICKABLE ROWS) */
-									<div className="overflow-x-auto">
-										<Table>
-											<TableHeader>
-												<TableRow className="hover:bg-transparent border-b">
-													<TableHead className="w-12 font-bold px-4">No</TableHead>
-													<TableHead className="font-bold">Siswa</TableHead>
-													<TableHead className="text-center font-bold">NIS</TableHead>
-													<TableHead className="text-center font-bold">H</TableHead>
-													<TableHead className="text-center font-bold">S</TableHead>
-													<TableHead className="text-center font-bold">I</TableHead>
-													<TableHead className="text-center font-bold">A</TableHead>
-													<TableHead className="text-center font-bold">Total</TableHead>
-													<TableHead className="text-center font-bold">%</TableHead>
-													<TableHead className="w-12"></TableHead>
-												</TableRow>
-											</TableHeader>
-											<TableBody>
-												{studentStatsList.map((st, index) => (
-													<TableRow
-														key={st.stId}
-														onClick={() => setSelectedStudentId(st.stId)}
-														className="hover:bg-muted/80 cursor-pointer group transition-colors border-b"
-													>
-														<TableCell className="px-4 text-xs font-medium text-muted-foreground">{index + 1}</TableCell>
-														<TableCell>
-															<div className="flex items-center gap-3">
-																<Avatar className="w-9 h-9 border-2 border-background shadow-xs">
-																	<AvatarImage src={st.photo} alt={st.studentName} />
-																	<AvatarFallback className="bg-primary/10 text-primary font-bold">
-																		{st.studentName.charAt(0).toUpperCase()}
-																	</AvatarFallback>
-																</Avatar>
-																<span className="font-semibold text-sm group-hover:text-primary transition-colors">
-																	{st.studentName}
-																</span>
-															</div>
-														</TableCell>
-														<TableCell className="text-center font-mono text-xs text-muted-foreground">{st.nipd}</TableCell>
-														<TableCell className="text-center">
-															<Badge className="bg-emerald-500 hover:bg-emerald-600 shadow-2xs border-none font-bold">
-																{st.countH}
-															</Badge>
-														</TableCell>
-														<TableCell className="text-center">
-															<Badge className="bg-sky-500 hover:bg-sky-600 shadow-2xs border-none font-bold">
-																{st.countS}
-															</Badge>
-														</TableCell>
-														<TableCell className="text-center">
-															<Badge className="bg-indigo-500 hover:bg-indigo-600 shadow-2xs border-none font-bold">
-																{st.countI}
-															</Badge>
-														</TableCell>
-														<TableCell className="text-center">
-															<Badge className="bg-amber-500 hover:bg-amber-600 shadow-2xs border-none font-bold">
-																{st.countA}
-															</Badge>
-														</TableCell>
-														<TableCell className="font-bold text-center text-xs">{st.totalDays}</TableCell>
-														<TableCell className="text-center">{getAttendanceRateBadge(st.rate)}</TableCell>
-														<TableCell className="text-center">
-															<IconChevronRight className="w-4 h-4 text-muted-foreground group-hover:text-primary transition-colors" />
-														</TableCell>
-													</TableRow>
-												))}
-											</TableBody>
-										</Table>
-									</div>
-								) : (
-									/* MATRIX TABLE VIEW (DAYS 1-31) */
-									<div className="overflow-x-auto w-full">
-										<Table className="min-w-max border-collapse">
-											<TableHeader>
-												<TableRow className="hover:bg-transparent border-b">
-													<TableHead className="w-12 text-center font-bold px-3">No</TableHead>
-													<TableHead className="min-w-48 font-bold px-4">Siswa</TableHead>
-													<TableHead className="text-center font-bold px-3">NIS</TableHead>
-													{daysArray.map((day) => (
-														<TableHead key={day} className="text-center w-9 min-w-9 px-1 font-semibold text-xs">
-															{day}
-														</TableHead>
-													))}
-													<TableHead className="text-center font-bold text-emerald-600 w-10 px-2">H</TableHead>
-													<TableHead className="text-center font-bold text-amber-600 w-10 px-2">S</TableHead>
-													<TableHead className="text-center font-bold text-blue-600 w-10 px-2">I</TableHead>
-													<TableHead className="text-center font-bold text-red-600 w-10 px-2">A</TableHead>
-												</TableRow>
-											</TableHeader>
-											<TableBody>
-												{studentStatsList.map((st, idx) => (
-													<TableRow
-														key={st.stId}
-														onClick={() => setSelectedStudentId(st.stId)}
-														className="border-b hover:bg-muted/40 cursor-pointer transition-colors"
-													>
-														<TableCell className="text-center font-medium text-xs text-muted-foreground w-12 px-3">
-															{idx + 1}
-														</TableCell>
-														<TableCell className="font-semibold text-xs min-w-48 px-4">
-															{st.studentName}
-														</TableCell>
-														<TableCell className="text-center text-xs font-mono text-muted-foreground px-3">
-															{st.nipd}
-														</TableCell>
-														{daysArray.map((day) => {
-															const status = st.dayStatusMap[day];
-															let badgeColor = "bg-muted text-muted-foreground/30";
-
-															if (status === "h") badgeColor = "bg-emerald-500 text-white font-bold";
-															else if (status === "s") badgeColor = "bg-amber-500 text-white font-bold";
-															else if (status === "i") badgeColor = "bg-blue-500 text-white font-bold";
-															else if (status === "a") badgeColor = "bg-red-500 text-white font-bold";
-
-															return (
-																<TableCell key={day} className="text-center p-1 text-xs">
-																	{status ? (
-																		<span
-																			className={`inline-flex items-center justify-center w-6 h-6 rounded-md text-[10px] ${badgeColor}`}
-																		>
-																			{status.toUpperCase()}
-																		</span>
-																	) : (
-																		<span className="text-muted-foreground/20 text-[10px]">•</span>
-																	)}
-																</TableCell>
-															);
-														})}
-														<TableCell className="text-center font-bold text-xs text-emerald-600 bg-emerald-50/40 dark:bg-emerald-950/20">
-															{st.countH}
-														</TableCell>
-														<TableCell className="text-center font-bold text-xs text-amber-600 bg-amber-50/40 dark:bg-amber-950/20">
-															{st.countS}
-														</TableCell>
-														<TableCell className="text-center font-bold text-xs text-blue-600 bg-blue-50/40 dark:bg-blue-950/20">
-															{st.countI}
-														</TableCell>
-														<TableCell className="text-center font-bold text-xs text-red-600 bg-red-50/40 dark:bg-red-950/20">
-															{st.countA}
-														</TableCell>
-													</TableRow>
-												))}
-											</TableBody>
-										</Table>
-									</div>
-								)}
-							</CardContent>
-						</Card>
 					</div>
 				)}
 			</div>
@@ -656,136 +132,100 @@ export default function PresensiSiswaRekapPage() {
 	);
 }
 
-// Attendance Calendar component matching /rekap/bulanan/[rombelId]/siswa/[studentId]
-function AttendanceCalendar({
+function RombelApelCard({
+	rombel,
 	month,
 	year,
-	dayStatusMap,
 }: {
+	rombel: any;
 	month: number;
 	year: number;
-	dayStatusMap: Record<number, string>;
 }) {
-	const dayNames = ["Min", "Sen", "Sel", "Rab", "Kam", "Jum", "Sab"];
+	const { data: monthlyData } = useMonthlyApelAttendanceQuery({
+		rombel_id: rombel.id,
+		month,
+		year,
+	});
 
-	const calendarDays = useMemo(() => {
-		const firstDay = new Date(year, month - 1, 1);
-		const lastDay = new Date(year, month, 0);
-		const daysInMonth = lastDay.getDate();
-		const startDayOfWeek = firstDay.getDay(); // 0 = Sunday
+	const studentAttendanceMap = monthlyData?.attendances ?? {};
+	const studentKeys = Object.keys(studentAttendanceMap);
 
-		const days: { key: string; day: number | null }[] = [];
+	const counts = useMemo(() => {
+		let countH = 0;
+		let countS = 0;
+		let countI = 0;
+		let countA = 0;
 
-		for (let i = 0; i < startDayOfWeek; i++) {
-			days.push({ key: `empty-${i}`, day: null });
-		}
+		studentKeys.forEach((stId) => {
+			const records = studentAttendanceMap[stId] || [];
+			records.forEach((rec) => {
+				const status = (rec.attendance_status || "h").toLowerCase();
+				if (status === "h") countH++;
+				else if (status === "s") countS++;
+				else if (status === "i") countI++;
+				else if (status === "a") countA++;
+			});
+		});
 
-		for (let day = 1; day <= daysInMonth; day++) {
-			days.push({ key: `day-${day}`, day });
-		}
+		return { countH, countS, countI, countA };
+	}, [studentKeys, studentAttendanceMap]);
 
-		return days;
-	}, [month, year]);
+	const waliName =
+		typeof rombel.wali_kelas === "object"
+			? rombel.wali_kelas?.fullname
+			: typeof rombel.wali_kelas === "string"
+				? rombel.wali_kelas
+				: "Wali Kelas";
 
-	const getStatusColor = (status: string | undefined) => {
-		if (!status) return "bg-muted border text-muted-foreground";
-		switch (status.toLowerCase()) {
-			case "h":
-				return "bg-emerald-500 text-white font-bold";
-			case "s":
-				return "bg-sky-500 text-white font-bold";
-			case "i":
-				return "bg-indigo-500 text-white font-bold";
-			case "a":
-				return "bg-amber-500 text-white font-bold";
-			default:
-				return "bg-muted border text-muted-foreground";
-		}
-	};
-
-	const getStatusLabel = (status: string | undefined) => {
-		if (!status) return "Tidak ada data";
-		switch (status.toLowerCase()) {
-			case "h":
-				return "Hadir";
-			case "s":
-				return "Sakit";
-			case "i":
-				return "Izin";
-			case "a":
-				return "Alpa";
-			default:
-				return status;
-		}
-	};
+	const jurusanNama =
+		typeof rombel.jurusan === "object"
+			? rombel.jurusan?.nama || rombel.jurusan?.kode || "Presensi Apel Pagi"
+			: rombel.jurusan || "Presensi Apel Pagi";
 
 	return (
-		<TooltipProvider>
-			<div className="w-full">
-				<div className="gap-1 grid grid-cols-7 mb-2">
-					{dayNames.map((day) => (
-						<div key={day} className="p-2 font-medium text-muted-foreground text-sm text-center">
-							{day}
+		<Link href={`/presensi-siswa/rekap/${rombel.id}?month=${month}&year=${year}`}>
+			<Card className="hover:shadow-md h-full hover:scale-95 transition duration-300 cursor-pointer border">
+				<CardHeader className="pb-2">
+					<CardTitle className="flex justify-between items-center text-lg">
+						<span>{rombel.nama}</span>
+						<Badge variant="secondary" className="text-xs font-normal">
+							Tingkat {rombel.tingkat_kelas}
+						</Badge>
+					</CardTitle>
+					<CardDescription>{jurusanNama}</CardDescription>
+				</CardHeader>
+				<CardContent>
+					<div className="space-y-2">
+						<div className="flex items-center gap-2 text-muted-foreground text-xs">
+							<User className="w-4 h-4" />
+							<span>{waliName}</span>
 						</div>
-					))}
-				</div>
+						<div className="flex items-center gap-2 text-muted-foreground text-xs">
+							<IconUsers className="w-4 h-4" />
+							<span>{studentKeys.length > 0 ? `${studentKeys.length} siswa` : "Buka rekap kelas"}</span>
+						</div>
 
-				<div className="gap-2 grid grid-cols-7">
-					{calendarDays.map(({ key, day }) => {
-						if (day === null) {
-							return <div key={key} className="p-2" />;
-						}
-
-						const status = dayStatusMap[day];
-						const statusColor = getStatusColor(status);
-						const statusLabel = getStatusLabel(status);
-
-						return (
-							<Tooltip key={key}>
-								<TooltipTrigger asChild>
-									<div
-										className={`p-2 md:p-3 rounded-lg text-center cursor-default duration-300 transition-all hover:scale-95 ${statusColor}`}
-									>
-										<div className="font-medium text-sm md:text-base">{day}</div>
-										<div className="hidden md:block opacity-90 mt-1 text-xs">
-											{status ? status.toUpperCase() : "-"}
-										</div>
-									</div>
-								</TooltipTrigger>
-								<TooltipContent>
-									<p className="text-xs">
-										Tanggal {day}: <span className="font-semibold">{statusLabel}</span>
-									</p>
-								</TooltipContent>
-							</Tooltip>
-						);
-					})}
-				</div>
-
-				{/* Legend */}
-				<div className="flex flex-wrap gap-4 mt-6 pt-4 border-t text-xs">
-					<div className="flex items-center gap-2">
-						<div className="bg-emerald-500 rounded w-4 h-4" />
-						<span>Hadir (H)</span>
+						{/* Attendance Stats Badges on Card */}
+						<div className="mt-3 pt-2 border-t">
+							<div className="mb-2 font-medium text-xs text-muted-foreground">Hitungan Presensi Apel:</div>
+							<div className="grid grid-cols-4 gap-1">
+								<Badge variant="outline" className="flex flex-col items-center p-1 border-emerald-300 dark:border-emerald-700">
+									<span className="font-bold text-emerald-600 text-xs">H: {counts.countH}</span>
+								</Badge>
+								<Badge variant="outline" className="flex flex-col items-center p-1 border-sky-300 dark:border-sky-700">
+									<span className="font-bold text-sky-600 text-xs">S: {counts.countS}</span>
+								</Badge>
+								<Badge variant="outline" className="flex flex-col items-center p-1 border-indigo-300 dark:border-indigo-700">
+									<span className="font-bold text-indigo-600 text-xs">I: {counts.countI}</span>
+								</Badge>
+								<Badge variant="outline" className="flex flex-col items-center p-1 border-amber-300 dark:border-amber-700">
+									<span className="font-bold text-amber-600 text-xs">A: {counts.countA}</span>
+								</Badge>
+							</div>
+						</div>
 					</div>
-					<div className="flex items-center gap-2">
-						<div className="bg-sky-500 rounded w-4 h-4" />
-						<span>Sakit (S)</span>
-					</div>
-					<div className="flex items-center gap-2">
-						<div className="bg-indigo-500 rounded w-4 h-4" />
-						<span>Izin (I)</span>
-					</div>
-					<div className="flex items-center gap-2">
-						<div className="bg-amber-500 rounded w-4 h-4" />
-						<span>Alpa (A)</span>
-					</div>
-					<div className="flex items-center gap-2">
-						<div className="bg-muted border rounded w-4 h-4" />
-						<span>Tidak Ada Data</span>
-					</div>
-				</div>
-			</div>
-		</TooltipProvider>
+				</CardContent>
+			</Card>
+		</Link>
 	);
 }
